+++ src/components/activation-displays/action-renderer.tsx
@@ -1,4 +1,4 @@
-import { ActionType, Activation } from '@/models/equipment'
+import { ActionType, Activation } from '@/models/activation'
 import Image from 'next/image'
 import { ReactNode } from 'react'
