+++ src/models/equipment.tsx
@@ -1,5 +1,5 @@
+import { Activation } from './activation'
 import { EntityModel } from './entity-model'
-import { AreaDefinition, RangeDefinition, Targets } from './spell'
 
 export type EquipmentCategory =
   | 'Adventuring Gear'
@@ -15,14 +15,6 @@ export type EquipmentUsage =
   | 'affixed to armor'
   | 'held in 2 hands'
 export type Rarity = 'uncommon' | 'rare'
-export type ActionName = 'Interact' | 'Cast a Spell'
-export type ActionType =
-  | 'one'
-  | 'two'
-  | 'three'
-  | 'reaction'
-  | 'free'
-  | '10 minutes'
 export type CurrencyType = 'gp'
 
 export interface Equipment extends EntityModel {
@@ -46,31 +38,6 @@ export interface EquipmentWithVariants extends EntityModel {
   activation?: Activation
   types: EquipmentVariantType[]
   rarity?: Rarity
-}
-
-export interface Activation {
-  num_actions: ActionType
-  action?: ActionName
-  traits?: string[]
-  frequency?: string // should this be an enum?
-  trigger?: string
-  effect?: Effect
-  targets?: Targets
-  range?: RangeDefinition
-  area?: AreaDefinition
-  override_label?: string
-}
-
-export interface Effect {
-  description: string
-  saving_throw?: SavingThrow
-}
-
-export interface SavingThrow {
-  critical_success?: string
-  success?: string
-  failure?: string
-  critical_failure?: string
 }
 
 export interface Currency {
