+++ src/models/spell.tsx
@@ -1,11 +1,7 @@
+import { Activation, SavingThrowType } from './activation'
 import { EntityModel } from './entity-model'
-import { Activation } from './equipment'
 
 export type Tradition = 'arcane' | 'primal' | 'divine' | 'occult'
-export type RangeUnit = 'feet'
-export type AreaUnit = 'emanataion'
-export type Targets = '1 or 2 creatures'
-export type SavingThrowType = 'basic Reflex'
 export type HeightenedType = 'formula' | 'explicit'
 
 export interface Spell extends EntityModel {
@@ -16,16 +12,6 @@ export interface Spell extends EntityModel {
   activation: Activation
   saving_throw?: SavingThrowType
   heightened: HeightenedDefinition
-}
-
-export interface RangeDefinition {
-  value: number
-  unit: RangeUnit
-}
-
-export interface AreaDefinition {
-  value: number
-  unit: AreaUnit
 }
 
 export interface HeightenedDefinition {
